Adds circle_ci config.yml, testing with python2 and 3
We use two off-the-shelf docker images, one for py2 and one for py3. We
then install the requirements on top on that (rather than have dedicated
docker images for sphinx_pyrelease).

As a sanity check, in each job, we print the python version that will be
under test.

This will require some fix-ups to code that explictly runs python3.